feat: 라이벌 없을 때 페이지 진입 대신 모달 표시
