+++ apps/mac/src/renderer/src/app/layouts/tab/TabLayout.types.ts
@@ -1,4 +1,5 @@
 export interface Tab {
   name: string;
   url: string;
+  onSelect?: () => boolean | void;
 }
