+++ apps/mac/src/renderer/src/pages/competition/index.ts
@@ -1,2 +1,3 @@
+export { CompetitionTabs } from "./CompetitionTabs";
 export { MyCompetitionPage } from "./my/MyCompetitionPage";
 export { RivalCompetitionPage } from "./rival/RivalCompetitionPage";
